refactor(all): 폴더 구조 변경
UI를 담당하는 로직인 InputView.js, OutputView.js를 view 폴더에 별도로 분류한다.
핵심 로직과 UI 로직을 분리하기 위함이다.
