+++ src/view/OutputView.js
@@ -1,6 +1,6 @@
 const { Console } = require("@woowacourse/mission-utils");
-const Player = require("./Player");
-const { MESSAGE } = require("./constant/Constants");
+const Player = require("../Player");
+const { MESSAGE } = require("../constant/Constants");
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
